Add own ncore creds

+++ frontend/src/Manage/Admin/DeleteUser/DeleteUser.jsx
@@ -1,5 +1,5 @@
 import { Form, Button } from 'react-bootstrap'
-import { useState, useContext } from 'react'
+import { useState, useContext, useEffect } from 'react'
 import { manageContext } from '../../Manage'
 import './DeleteUser.css'
 import { useTranslation } from 'react-i18next'
@@ -7,30 +7,36 @@ import { LOGIN_PAGE, redirectToPage } from '../../../util'
 
 export const DeleteUser = ({ fetchUsers, users }) => {
   const { t } = useTranslation()
-  const [selectedUser, setSelectedUser] = useState(users[0]?.email)
+  const [selectedUser, setSelectedUser] = useState('')
   const { setToastData } = useContext(manageContext)
+
+  useEffect(() => {
+    setSelectedUser(users[0]?.email)
+  }, [users])
 
   const handleDelete = async (event) => {
     event.preventDefault()
-    const id = users.filter(user => user.email === selectedUser)[0].id
-    try {
-      const resp = await fetch(`/api/users/${id}/`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json'
+    if (window.confirm(t('DELETE_USER_CONFIRM'))) {
+      const id = users.filter(user => user.email === selectedUser)[0].id
+      try {
+        const resp = await fetch(`/api/users/${id}/`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json'
+          }
+        })
+        if (resp.status === 200) {
+          setToastData({ message: t('USER_DELETE_SUCCESS'), type: 'success' })
+        } else if (resp.status === 401) {
+          redirectToPage(LOGIN_PAGE)
+        } else {
+          setToastData({ message: t('USER_DELETE_ERROR'), type: 'danger' })
         }
-      })
-      if (resp.status === 200) {
-        setToastData({ message: t('USER_DELETE_SUCCESS'), type: 'success' })
-      } else if (resp.status === 401) {
-        redirectToPage(LOGIN_PAGE)
-      } else {
-        setToastData({ message: t('USER_DELETE_ERROR'), type: 'danger' })
+      } catch (error) {
+        setToastData({ message: t('UNEXPECTED_ERROR'), type: 'danger' })
       }
-    } catch (error) {
-      setToastData({ message: t('UNEXPECTED_ERROR'), type: 'danger' })
+      fetchUsers()
     }
-    fetchUsers()
   }
 
   return (
